requirements.txt
autopep8==1.6.0
click==8.0.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
SQLAlchemy==1.4.32
tweepy==4.4.0
twitchio==2.2.0
python-dotenv==0.13.0
unidecode==1.2.0
